Refactored file into module to for better structure
